pyproject.toml
[project]
name = "data-product-hub"
version = "0.1.0"
description = "Config-driven data product hub with OData-style query API and Kubernetes operator."
readme = "README.md"
requires-python = ">=3.10"

authors = [
  { name = "Your Name" }
]

dependencies = [
  # --- Engine Runtime ---
  "fastapi>=0.110.0",
  "uvicorn>=0.28.0",
  "pydantic>=2.0",
  "pandas>=2.0",
  "duckdb>=0.10.0",
  "pyarrow>=14.0.1",
  "pyyaml>=6.0",
  "pyjwt[crypto]>=2.8.0",

  # --- Operator Runtime ---
  "kopf>=1.37.0",
  "kubernetes>=28.1.0",
  "requests>=2.31.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "httpx>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
